Address UI editor preview review findings
