+++ clients/silencer/src/net/ui_editor_preview_document.h
@@ -1,7 +1,7 @@
 #ifndef UI_EDITOR_PREVIEW_DOCUMENT_H
 #define UI_EDITOR_PREVIEW_DOCUMENT_H
 
-#include "ui_editor_preview_screen.h"
+#include "ui_editor_preview_model.h"
 
 #include "nlohmann/json.hpp"
 
@@ -10,7 +10,7 @@
 namespace silencer::net {
 
 bool ParseUiEditorPreviewDocument(const nlohmann::json& json,
-                                  silencer::client_ui::UiEditorPreviewDocument& document,
+                                  silencer::ui::UiEditorPreviewDocument& document,
                                   std::string& error);
 
 }  // namespace silencer::net
