+++ web/admin/app/ui-editor/components/Canvas.tsx
@@ -1,0 +1,141 @@
+import {
+  PALETTE_NODE_KINDS,
+  canHaveChildren,
+  findNode,
+  type UiDocument,
+  type UiNodeKind,
+} from '../../../lib/ui-layout';
+import { type ClientPreviewElement, type ClientPreviewState } from '../useClientPreview';
+import { PreviewNode } from './BrowserPreview';
+
+interface CanvasProps {
+  document: UiDocument;
+  selectedId: string;
+  zoom: number;
+  clientPreview: ClientPreviewState;
+  onSelect: (id: string) => void;
+  onDropNode: (targetId: string, kind: UiNodeKind) => void;
+}
+
+export function Canvas({ document, selectedId, zoom, clientPreview, onSelect, onDropNode }: CanvasProps) {
+  const live = clientPreview.status === 'live' && clientPreview.screenshot;
+  return (
+    <div className="min-h-0 flex-1 overflow-auto p-6">
+      <div
+        className="relative mx-auto shadow-[0_0_0_1px_rgba(86,94,111,0.8),0_24px_80px_rgba(0,0,0,0.65)]"
+        style={{
+          width: Math.round(document.viewport.width * zoom),
+          height: Math.round(document.viewport.height * zoom),
+        }}
+      >
+        <div
+          className="origin-top-left relative"
+          style={{
+            width: document.viewport.width,
+            height: document.viewport.height,
+            transform: `scale(${zoom})`,
+          }}
+        >
+          {live ? (
+            <>
+              <img
+                alt=""
+                src={clientPreview.screenshot}
+                className="absolute inset-0 h-full w-full select-none"
+                draggable={false}
+              />
+              <ClientPreviewOverlay
+                elements={clientPreview.elements}
+                document={document}
+                selectedId={selectedId}
+                onSelect={onSelect}
+                onDropNode={onDropNode}
+              />
+            </>
+          ) : (
+            <PreviewNode
+              node={document.root}
+              selectedId={selectedId}
+              rootViewport={document.viewport}
+              onSelect={onSelect}
+              onDropNode={onDropNode}
+            />
+          )}
+          <div className="absolute left-2 top-2 border border-game-border bg-game-bg/85 px-2 py-1 text-[10px] tracking-widest text-game-textDim">
+            {clientPreview.status === 'live' ? 'CLIENT LIVE' : clientPreview.status === 'syncing' ? 'SYNCING CLIENT' : 'BROWSER FALLBACK'}
+          </div>
+        </div>
+      </div>
+      {clientPreview.status === 'offline' && clientPreview.error && (
+        <div className="mx-auto mt-3 max-w-[720px] border border-game-danger/60 bg-game-bgCard px-3 py-2 text-[11px] tracking-wider text-game-danger">
+          {clientPreview.error}
+        </div>
+      )}
+    </div>
+  );
+}
+
+function ClientPreviewOverlay({ elements, document, selectedId, onSelect, onDropNode }: {
+  elements: ClientPreviewElement[];
+  document: UiDocument;
+  selectedId: string;
+  onSelect: (id: string) => void;
+  onDropNode: (targetId: string, kind: UiNodeKind) => void;
+}) {
+  return (
+    <div className="absolute inset-0">
+      {elements.map(element => (
+        <ClientPreviewOverlayTarget
+          key={`${element.id ?? 'anonymous'}-${element.source}-${element.kind ?? 'element'}`}
+          element={element}
+          document={document}
+          selectedId={selectedId}
+          onSelect={onSelect}
+          onDropNode={onDropNode}
+        />
+      ))}
+    </div>
+  );
+}
+
+function ClientPreviewOverlayTarget({ element, document, selectedId, onSelect, onDropNode }: {
+  element: ClientPreviewElement;
+  document: UiDocument;
+  selectedId: string;
+  onSelect: (id: string) => void;
+  onDropNode: (targetId: string, kind: UiNodeKind) => void;
+}) {
+  if (!element.id) return null;
+  const node = findNode(document.root, element.id);
+  if (!node) return null;
+  const selected = element.id === selectedId;
+  return (
+    <div
+      className="absolute"
+      title={element.label ?? element.id}
+      onClick={event => {
+        event.stopPropagation();
+        onSelect(element.id!);
+      }}
+      onDragOver={event => {
+        if (canHaveChildren(node.kind)) event.preventDefault();
+      }}
+      onDrop={event => {
+        if (!canHaveChildren(node.kind)) return;
+        const kind = event.dataTransfer.getData('application/silencer-ui-kind') as UiNodeKind;
+        if (!kind || !PALETTE_NODE_KINDS.includes(kind)) return;
+        event.preventDefault();
+        event.stopPropagation();
+        onDropNode(element.id!, kind);
+      }}
+      style={{
+        left: element.x,
+        top: element.y,
+        width: element.w,
+        height: element.h,
+        outline: selected ? '2px solid #f59e0b' : '1px solid rgba(245, 158, 11, 0.22)',
+        outlineOffset: selected ? 2 : 0,
+      }}
+    />
+  );
+}
