+++ web/admin/app/ui-editor/ui-editor-utils.ts
@@ -1,0 +1,17 @@
+export function slugify(value: string): string {
+  return value.trim().toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/(^-|-$)/g, '') || 'unnamed';
+}
+
+export function normalizeColor(value: string): string {
+  return /^#[0-9a-f]{6}$/i.test(value) ? value : '#000000';
+}
+
+export function downloadText(filename: string, text: string, mime: string) {
+  const blob = new Blob([text], { type: mime });
+  const url = URL.createObjectURL(blob);
+  const link = window.document.createElement('a');
+  link.href = url;
+  link.download = filename;
+  link.click();
+  URL.revokeObjectURL(url);
+}
